Change sql action package name
