+++ stdlib/ballerina-builtin/src/main/java/org/ballerinalang/nativeimpl/sql/Constants.java
@@ -15,7 +15,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.ballerinalang.nativeimpl.actions.data.sql;
+package org.ballerinalang.nativeimpl.sql;
 
 /**
  * Constants for SQL Data Connectors.
