+++ stdlib/ballerina-builtin/src/main/java/org/ballerinalang/nativeimpl/sql/SQLDatasource.java
@@ -15,7 +15,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.ballerinalang.nativeimpl.actions.data.sql;
+package org.ballerinalang.nativeimpl.sql;
 
 import com.zaxxer.hikari.HikariConfig;
 import com.zaxxer.hikari.HikariDataSource;
