+++ stdlib/ballerina-builtin/src/main/java/org/ballerinalang/nativeimpl/sql/actions/BatchUpdate.java
@@ -15,15 +15,14 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.ballerinalang.nativeimpl.actions.data.sql.actions;
+package org.ballerinalang.nativeimpl.sql.actions;
 
 import org.ballerinalang.bre.Context;
-import org.ballerinalang.model.types.BStructType;
 import org.ballerinalang.model.types.TypeKind;
 import org.ballerinalang.model.values.BRefValueArray;
 import org.ballerinalang.model.values.BStruct;
-import org.ballerinalang.nativeimpl.actions.data.sql.Constants;
-import org.ballerinalang.nativeimpl.actions.data.sql.SQLDatasource;
+import org.ballerinalang.nativeimpl.sql.Constants;
+import org.ballerinalang.nativeimpl.sql.SQLDatasource;
 import org.ballerinalang.natives.annotations.Argument;
 import org.ballerinalang.natives.annotations.BallerinaFunction;
 import org.ballerinalang.natives.annotations.Receiver;
@@ -36,34 +35,38 @@ import static org.ballerinalang.util.observability.ObservabilityConstants.TAG_KE
 import static org.ballerinalang.util.observability.ObservabilityConstants.TAG_KEY_DB_TYPE;
 
 /**
- * {@code Call} is the Call action implementation of the SQL Connector.
+ * {@code BatchUpdate} is the Batch update action implementation of the SQL Connector.
  *
- * @since 0.8.0
+ * @since 0.8.6
  */
 @BallerinaFunction(
         orgName = "ballerina", packageName = "sql",
-        functionName = "call",
-        receiver = @Receiver(type = TypeKind.STRUCT, structType = "ClientConnector"),
+        functionName = "batchUpdate",
+        receiver = @Receiver(type = TypeKind.STRUCT,
+                             structType = "ClientConnector",
+                             structPackage = "ballerina.sql"),
         args = {
+                @Argument(name = "client", type = TypeKind.STRUCT),
                 @Argument(name = "sqlQuery", type = TypeKind.STRING),
-                @Argument(name = "parameters", type = TypeKind.ARRAY, elementType = TypeKind.STRUCT,
+                @Argument(name = "parameters",
+                          type = TypeKind.ARRAY,
+                          elementType = TypeKind.STRUCT,
+                          arrayDimensions = 2,
                           structType = "Parameter")
         },
         returnType = {
-                @ReturnType(type = TypeKind.TABLE),
+                @ReturnType(type = TypeKind.ARRAY, elementType = TypeKind.INT),
                 @ReturnType(type = TypeKind.STRUCT, structType = "SQLConnectorError",
                             structPackage = "ballerina.sql")
         }
 )
-public class Call extends AbstractSQLAction {
-
+public class BatchUpdate extends AbstractSQLAction {
     @Override
     public void execute(Context context) {
         try {
             BStruct bConnector = (BStruct) context.getRefArgument(0);
             String query = context.getStringArgument(0);
             BRefValueArray parameters = (BRefValueArray) context.getNullableRefArgument(1);
-            BStructType structType = getStructType(context);
             SQLDatasource datasource = (SQLDatasource) bConnector.getNativeData(Constants.CLIENT_CONNECTOR);
 
             ObserverContext observerContext = ObservabilityUtils.getCurrentContext(context.
@@ -71,7 +74,7 @@ public class Call extends AbstractSQLAction {
             observerContext.addTag(TAG_KEY_DB_STATEMENT, query);
             observerContext.addTag(TAG_KEY_DB_TYPE, TAG_DB_TYPE_SQL);
 
-            executeProcedure(context, datasource, query, parameters, structType);
+            executeBatchUpdate(context, datasource, query, parameters);
         } catch (Throwable e) {
             context.setReturnValues(SQLDatasourceUtils.getSQLConnectorError(context, e));
             SQLDatasourceUtils.handleErrorOnTransaction(context);
