+++ tests/ballerina-test/src/test/resources/test-src/taintchecking/connectors/sql-select-untainted-query-tainted-return-negative.bal
@@ -1,4 +1,4 @@
-import ballerina/data.sql;
+import ballerina/sql;
 
 struct Employee {
     int id;
