+++ tests/ballerina-test/src/test/resources/test-src/types/table/table-iteration.bal
@@ -1,4 +1,4 @@
-import ballerina/data.sql;
+import ballerina/sql;
 
 struct Person {
     int id;
