+++ tests/ballerina-test/src/test/resources/test-src/types/table/table-test-helper.bal
@@ -1,4 +1,4 @@
-import ballerina/data.sql;
+import ballerina/sql;
 
 struct ResultCount {
     int COUNTVAL;
