+++ tests/ballerina-test/src/test/resources/test-src/types/table/table-type.bal
@@ -1,4 +1,4 @@
-import ballerina/data.sql;
+import ballerina/sql;
 import ballerina/io;
 import ballerina/time;
